requirements.txt
# pytorch 
--extra-index-url https://download.pytorch.org/whl/cu113
# transformers
torch~=1.0 
transformers~=4.20
# hydra
hydra-core~=1.2
# torch text
torchtext~=0.12 
# Config manager
# hydra
hydra-core~=1.2
# progress bar
tqdm~=4.64





